Add functionality to Start component
The start button when clicked calls the startGame method in the App
component. When run, the method will call the randomButtonGenerator
method. If the buttonPattern state currently has a length of zero it
will create the first number. If the length is greater than zero, a
parameter is passed which will clear the array before generating
the number (acting as a reset).