Add computerPlayButtonPattern component
This component will play through the pattern stored in state. It will
apply the active class for the respective button as well as play the
sound for it. After a second it will remove the active class.

This method uses recursion to iterate over the entire length of the
buttonPattern state. Additionally, the sound element was added to each
button components and is triggered on mouse down events.