Create basic game playability
A number of things have been added and changed to allow the basic
functions of playing the game.

* ColorPlayButtons component was updated to only allow buttons to be
clicked if the game is on and it is the player's turn.
* The toggleGamePower method was updated to reset the state to the
initial state if the game is turned off.
* A timeout was added to call the computerPlayButtonPattern method after
a new random button is generated.
* After the computer plays the button pattern it changes isPlayersTurn
to true and resets the playerCopyPattern to a blank array.
* A basic playerSelectButton method was created that checks the player
pattern against the computer random pattern. If the pattern is entered
correctly the game advances.