Fix bug with computerPlayButtonPattern
There was a bug if the user turned the game off while the computer
was playing the button pattern. This was due to turning the game off
erasing the buttonPattern from state. In order to resolve, a logical
test was added to this method that confirms the game is on for each
recursion.